+++ ContinuousStepper_Timer1.h
@@ -1,14 +1,22 @@
 #pragma once
 
+#include <TimerOne.h> // https://github.com/PaulStoffregen/TimerOne
+
 #include "ContinuousStepper.h"
 
-class ContinuousStepper_Timer1 {
+template <typename TTimer>
+class ContinuousStepper_Timer {
 public:
   using float_t = ContinuousStepper::float_t;
   using time_t = ContinuousStepper::time_t;
   using pin_t = ContinuousStepper::pin_t;
 
-  ContinuousStepper_Timer1(pin_t stepPin, pin_t dirPin, pin_t enablePin = ContinuousStepper::NULL_PIN);
+  ContinuousStepper_Timer(pin_t stepPin, pin_t dirPin, pin_t enablePin = ContinuousStepper::NULL_PIN) {
+    _stepper = new ContinuousStepper(stepPin, dirPin, enablePin);
+    _timer.attachInterrupt(interruptHandler);
+    _timerPeriod = _stepper->interval();
+    _timer.initialize(_timerPeriod);
+  }
 
   void powerOn() {
     _stepper->powerOn();
@@ -51,9 +59,35 @@ public:
   }
 
 private:
-  static void configureTimer();
-  static void interruptHandler();
+  static void configureTimer() {
+    if (_timerPeriod == _stepper->interval())
+      return;
+
+    _timer.stop();
+    _timerPeriod = _stepper->interval();
+    if (_timerPeriod) {
+      _timer.setPeriod(_timerPeriod);
+      _timer.start();
+    }
+  }
+
+  static void interruptHandler() {
+    _stepper->loop();
+    configureTimer();
+  }
 
   static ContinuousStepper *_stepper;
   static time_t _timerPeriod;
+  static TTimer _timer;
 };
+
+template <typename TTimer>
+ContinuousStepper *ContinuousStepper_Timer<TTimer>::_stepper;
+
+template <typename TTimer>
+ContinuousStepper::time_t ContinuousStepper_Timer<TTimer>::_timerPeriod;
+
+template <typename TTimer>
+TTimer ContinuousStepper_Timer<TTimer>::_timer;
+
+typedef ContinuousStepper_Timer<TimerOne> ContinuousStepper_Timer1;